Moved updateOrRemoveSchedule to engage mutations
